feat(sign_up_screen.dart): add confirm password field and validation
Add a confirm password field to the sign up screen and a validation to ensure that the password and confirm password fields match. Also, add a new validator function to the validators.dart file to validate that two strings are the same.

feat(perso_text_field.dart): add confirm password controller

Add a confirm password controller to the PersoTextField widget to allow for the validation of the confirm password field. Update the validator function to use the confirm password controller if it is present.